+++ frontend/src/metabase/core/components/AccordionList/AccordionList.styled.tsx
@@ -1,0 +1,5 @@
+import styled from "styled-components";
+
+export const AccordionListRoot = styled.div`
+  outline: none;
+`;
