+++ frontend/src/metabase/css/core/base.css
@@ -156,3 +156,7 @@ textarea {
 .no-outline * {
   outline: none;
 }
+
+.outline-none {
+  outline: none;
+}
